LayoutTests/imported/w3c/web-platform-tests/wasm/core/simd/simd_memory-multi.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds (simd_memory-multi.wast:5) assert_equals: expected false but got true
PASS #3 Reinitialize the default imports
FAIL #4 Test that WebAssembly compilation succeeds (simd_memory-multi.wast:5) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 33: Memory section has more than one memory, WebAssembly currently only allows zero or one at {loc} expected true got false
FAIL #5 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
simd_memory_multi_wast_js@http://web-platform.test:8800/wasm/core/js/simd/simd_memory-multi.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/simd/simd_memory-multi.wast.js:9:3 expected true got false
